Massive Bug fixes, optimizations, security hardening, testing, and other updates

+++ src/txdb.cpp
@@ -206,7 +206,7 @@ bool CCoinsViewDBCursor::GetValue(Coin &coin) const
     return pcursor->GetValue(coin);
 }
 
-unsigned int CCoinsViewDBCursor::GetValueSize() const
+size_t CCoinsViewDBCursor::GetValueSize() const
 {
     return pcursor->GetValueSize();
 }
